refactor: reorganized files into subfolders and renamed file to SupportedServiceType to be more expressive to what it is and does
