+++ ebms-send-in/src/main/kotlin/no/nav/emottak/ebms/plugin/Routes.kt
@@ -3,8 +3,8 @@ package no.nav.emottak.ebms.plugin
 import io.ktor.server.application.Application
 import io.ktor.server.routing.routing
 import io.micrometer.prometheus.PrometheusMeterRegistry
-import no.nav.emottak.ebms.fagmeldingRoutes
-import no.nav.emottak.ebms.healthcheckRoutes
+import no.nav.emottak.ebms.routes.fagmeldingRoutes
+import no.nav.emottak.ebms.routes.healthcheckRoutes
 
 fun Application.configureRoutes(prometheusMeterRegistry: PrometheusMeterRegistry) {
     routing {
