+++ ebms-send-in/src/main/kotlin/no/nav/emottak/ebms/routes/HealthcheckRoutes.kt
@@ -1,4 +1,4 @@
-package no.nav.emottak.ebms
+package no.nav.emottak.ebms.routes
 
 import io.ktor.server.response.respond
 import io.ktor.server.response.respondText
